V1 Design
This is our first design for the 10K COTS rocket, everything is subject to change. This design is based off the 98mm diameter m-motor. Please add your criticisms and potential ideas to this design. Another design will be made incorporating the 75mm m-motor later on, and a decision can be made on the motor on a later date.